requirements.txt
streamlit
streamlit-option-menu
ultralytics
tensorflow
numpy
pillow
opencv-python-headless
